+++ tests/testHDF5IO.cpp
@@ -32,6 +32,24 @@ std::string executablePath = "./reader_executable";
 using namespace AQNWB;
 namespace fs = std::filesystem;
 
+#ifdef H5_HAVE_ROS3_VFD
+TEST_CASE("HDF5IO; ROS3 mode", "[hdf5io]")
+{
+  std::string s3Url =
+      "https://dandiarchive.s3.amazonaws.com/blobs/fec/8a6/"
+      "fec8a690-2ece-4437-8877-8a002ff8bd8a";
+
+  SECTION("open with ros3 driver - region only")
+  {
+    IO::HDF5::HDF5IO hdf5io(s3Url);
+    Status status = hdf5io.open_s3("us-east-2");
+    REQUIRE(status == Status::Success);
+    REQUIRE(hdf5io.isOpen());
+    hdf5io.close();
+  }
+}
+#endif
+
 TEST_CASE("open - hdf5 file modes", "[hdf5io]")
 {
   const std::string fileName = getTestFilePath("test_open_modes.h5");